requirements.txt
google-cloud==0.32.0
numpy==1.14.2
pandas==0.22.0
scipy==1.0.0
scikit-learn==0.19.1
